requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
faiss-cpu
fastapi
langchain-huggingface
langchain-mistralai
scikit-learn
streamlit